requirements.txt
streamlit
pandas
numpy
matplotlib
scikit-learn
psycopg2
plotly
python-dotenv